support notifications from subversion to give feedback on what's happening in the working copy ;) this work for all subversion operations since they use the same system
here how it works :
the app asks to 'svn up' a file
kio_svn fires up, start to 'svn up'
during the progression of the update, notifications are sent back to kio_svn (from subversion library)
which forwards these to the kded module (via DCOP) which itself sends a DCOP signal with these informations.
So any listening app should receive the progress information.

I could not make the kioslave to emit directly the dcop signal, is this possible ?
(that would reduce this sytem by one DCOP message for each notification ... )

svn path=/trunk/kdesdk/kioslave/; revision=354822