Silence warnings about experimental 'when' and 'given'
Summary:
It seems that those keywords depends on an experimental feature and
this leads to a warning in recent version of Perl (where recent ==
few years already). So disable the warnings about the feature.

Test Plan:
No more warnings printed by the perl program (more visible when
the KIOSlave is invoked with kioclient5 cat perl:/...)

Reviewers: mpyne

Reviewed By: mpyne

Differential Revision: https://phabricator.kde.org/D5022